TINY-8755: Some improvements, moved a file from api when it was not api.
