+++ modules/tinymce/src/core/main/ts/dom/ElementUtils.ts
@@ -1,10 +1,17 @@
-import { Obj, Type } from '@ephox/katamari';
+import { Arr, Obj, Strings, Type } from '@ephox/katamari';
 
-import * as Bookmarks from '../../bookmark/Bookmarks';
-import * as NodeType from '../../dom/NodeType';
-import { attributeIsInternal } from '../../fmt/FormatUtils';
-import Tools from '../util/Tools';
-import DOMUtils from './DOMUtils';
+import Editor from '../api/Editor';
+import Tools from '../api/util/Tools';
+import * as Bookmarks from '../bookmark/Bookmarks';
+import * as NodeType from './NodeType';
+
+const internalAttributesPrefixes = [
+  'data-ephox-',
+  'data-mce-',
+  'data-alloy-',
+  'data-snooker-',
+  '_'
+];
 
 /**
  * Utility class for various element specific functions.
@@ -17,9 +24,12 @@ const each = Tools.each;
 
 export interface ElementUtils {
   readonly compare: (node1: Node, node2: Node) => boolean;
+  readonly attributeIsInternal: (attribute: string) => boolean;
 }
 
-const ElementUtils = (dom: DOMUtils): ElementUtils => {
+const ElementUtils = (editor: Editor): ElementUtils => {
+  const dom = editor.dom;
+  const internalAttributes = new Set<string>(editor.serializer.getTempAttrs());
   /**
    * Compares two nodes and checks if it's attributes and styles matches.
    * This doesn't compare classes as items since their order is significant.
@@ -112,8 +122,12 @@ const ElementUtils = (dom: DOMUtils): ElementUtils => {
     return !Bookmarks.isBookmarkNode(node1) && !Bookmarks.isBookmarkNode(node2);
   };
 
+  const attributeIsInternal = (attributeName: string): boolean =>
+    Arr.find(internalAttributesPrefixes, (value) => Strings.startsWith(attributeName, value)).isSome() || internalAttributes.has(attributeName);
+
   return {
-    compare
+    compare,
+    attributeIsInternal
   };
 };
